+++ src/NanoVNASaver/Controls/SerialControl.py
@@ -22,6 +22,8 @@ from time import sleep
 from PySide6 import QtWidgets
 from PySide6.QtCore import Signal
 
+from NanoVNASaver import NanoVNASaver
+
 from ..Hardware.Hardware import Interface, get_interfaces, get_VNA
 from .Control import Control
 
@@ -33,7 +35,7 @@ class SerialControl(Control):
     # true when serial port was connected and false when it was disconnected
     connected = Signal(bool)
 
-    def __init__(self, app: QtWidgets.QWidget):
+    def __init__(self, app: NanoVNASaver):
         super().__init__(app, "Serial port control")
 
         self.interface = Interface("serial", "none")
