+++ src/NanoVNASaver/Controls/SweepControl.py
@@ -20,6 +20,8 @@ import logging
 
 from PySide6 import QtCore, QtWidgets
 
+from NanoVNASaver import NanoVNASaver
+
 from ..Formatting import (
     format_frequency_short,
     format_frequency_sweep,
@@ -32,7 +34,7 @@ logger = logging.getLogger(__name__)
 
 
 class SweepControl(Control):
-    def __init__(self, app: QtWidgets.QWidget):
+    def __init__(self, app: NanoVNASaver):
         super().__init__(app, "Sweep control")
 
         line = QtWidgets.QFrame()
