+++ src/NanoVNASaver/__main__.py
@@ -30,8 +30,8 @@ import sys
 
 from PySide6 import QtWidgets
 
+from NanoVNASaver import NanoVNASaver
 from NanoVNASaver.About import INFO, VERSION
-from NanoVNASaver.NanoVNASaver import NanoVNASaver
 from NanoVNASaver.Touchstone import Touchstone
 from NanoVNASaver.utils import get_runtime_information
